pyproject.toml
[project]
name = "cordbeat"
version = "0.1.0"
description = "A local-first autonomous AI agent that thinks, acts, and stays by your side"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "websockets>=12.0",
    "pyyaml>=6.0",
    "chromadb>=1.0",
    "httpx>=0.27",
]

[project.scripts]
cordbeat = "cordbeat.main:cli"

[project.optional-dependencies]
discord = ["discord.py>=2.0"]
telegram = ["python-telegram-bot>=21.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "httpx>=0.27",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cordbeat"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
